Extract Parser, Ui, Storage classes
